Move FreeShippingConfig to 'checkout' package
